Add custom context menu
Add custom context menu for when its right click on stops in stop list or on stops in a route. For now all copying and 'show stop info' is functional, but edit, delete, remove, select are not functional.
Add global listener for mouse click and scroll to hide different popups. For now only context menu is added to that listener.