Add hiding of routesInfo and routeSelect on "blur"
Add routesInfo (that shows in what routes a stop is) and routeSelect dropdown list to the popUp handler that hides these pop ups when clicked or scrolled outsideof them.

Small style adjustments for the mentioned pop up.

Updated property name to 'mainApp' in 'ContextMenu' to have consistent naming with other modules.